Add jq to gosec producer container
gosec supports a configuration file, but the range of settings that can
be configured within it is limited; in particular, it lacks equivalents
for the -include, -exclude and -exclude-dir command line options. Add jq
to the gosec producer container, so that values for these options can be
specified in the configuration file and transformed into real command
line options by the Dracon producer.