Add -package-path option to npm audit producer
A code base scanned by Dracon can potentially contain multiple
JavaScript modules, but there is currently no way to disambiguate
between them by the time any vulnerable third-party dependencies in the
code base have been passed as issues to the enricher, making it
unnecessarily difficult to identify the locations of vulnerable
dependencies in the code base. To facilitate this, add a -package-path
option to the npm audit producer; when specified, the given value will
be prepended to the vulnerable dependency name in the respective issue's
Target field.