Add -append option to producers (#76)
To support incremental writing of producer output files, add support for
an -append option to the common producer library. When specified on the
command line, this appends issues from the file given by -in to the list
of issues in the protobuf stored in -out if -out already exists and is
non-empty, rather than overwriting the existing file entirely.