Fixed pipelineRefs updating
This was fixed by removing the `namespace` in the base kustomization as
this made Kustomize not recognise resources unless they were all in the
same namespace.

The tests are also working (and passing).

Next, I need to:
 - Verify that modifying components via patches is possible in
downstream Kustomizes (e.g. a golang-example wants to patch the
volumes/args in the gosec-producer component).
 - Play with NFS and a Tekton workspace.