Handle enrichment DB entries without CVEs (#141)
Entries in the enrichment DB added before the change to record CVEs have NULL values in the CVE column. This is not handled by the string type, resulting in errors trying to convert existing DB issues to enriched issues. This PR changes the CVE field in the DB issue struct to be sql.NullString, which does have handling for NULL values.