`yarn audit` producer: unconditionally write results file (#145)
The enricher always expects producers that run before it to output
results files, even if the producer found no issues to report. This
isn't the case with the `yarn audit` producer - a results file is only
written if at least one issue was found.

Write a results file regardless of how many issues were found by the
producer.

Fixes #144.